test(example): share e2e flow across integration_test + Patrol harnesses
Extract the OIDC conformance/smoke flow into a harness-agnostic
integration_test/shared_e2e.dart (its only coupling is a `pumpAndSettle`
callback), so the identical logic runs under both runners:

- integration_test/app_test.dart  -> thin testWidgets wrapper (linux/windows)
- patrol_test/app_test.dart       -> thin patrolTest wrapper (android/iOS/macOS/web)

No behavioral change to the conformance flow; the conformance/ + helpers.dart
support code is reused as-is. `dart analyze` on the example is clean (only a
pre-existing info), confirming the Patrol entrypoint compiles against patrol
4.5.0 and the shared import resolves.

CI wiring (activate patrol_cli from the fork, Playwright for web, swap job test
steps) follows in the next commit.
